+++ src/platform/linux/runtime.rs
@@ -122,6 +122,13 @@ where
     shared::block_on::<LinuxRuntime, F>(future)
 }
 
+pub fn try_block_on<F>(future: F) -> io::Result<F::Output>
+where
+    F: Future,
+{
+    shared::try_block_on::<LinuxRuntime, F>(future)
+}
+
 pub fn run_until_stalled() {
     shared::run_until_stalled::<LinuxRuntime>()
 }